pyproject.toml
[tool.poetry]
name = "base_graphql_with_user"
version = "0.1.0"
description = ""
authors = ["Francisco José Machado Junior <machado.fjunior@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
fastapi = "^0.115.1"
strawberry-graphql = "^0.266.1"
pydantic = { extras = ["email"], version = "^2.11.5" }
passlib = "^1.7.4"
python-jose = "^3.4.0"
sqlalchemy = "^2.0.40"
python-dotenv = "^1.1.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.9.1"
alembic = "^1.16.1"
uvicorn = "^0.34.2"
psycopg2-binary = "^2.9.10"
bcrypt = "<4.1.0"
orjson = "^3.10.18"
redis = "^6.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
httpx = "^0.28.1"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
faker = "^37.3.0"
pytest-cov = "^6.1.1"
docker = "^7.1.0"
pytest-order = "^1.3.0"
anyio = "^4.9.0"
pre-commit = "^4.2.0"
ruff = "^0.11.13"

[tool.poetry.scripts]
test = "run_tests:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --tb=short"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"] # Exemplos: erros, flake8, imports
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.pycodestyle]
max-line-length = 88
